+++ src/lib.rs
@@ -333,6 +333,10 @@ with `#[derive(Mmio)]`:
 - `#[mmio(no_ctors)]`: Omit the generation of constructor functions like
   `new_mmio_at` and `new_mmio`. This allows users to specify their own custom
   constructors, for example to constrain or check the allowed base addresses.
+- `#[mmio(const_ptr)]`: Pointer getter methods for array field are `const` now.
+  Requires Rust 1.83.0 or higher.
+- `#[mmio(const_inner)]`: Const getter methods for inner MMIO blocks. Requires Rust 1.83.0 or
+  higher.
 
 ### Field attributes
 
@@ -401,12 +405,12 @@ pub struct SharedInner<T>(T);
 
 impl<T> SharedInner<T> {
     #[doc(hidden)]
-    pub fn __new_internal(t: T) -> Self {
+    pub const fn __new_internal(t: T) -> Self {
         Self(t)
     }
 
     /// Get shared access to the value within
-    pub fn inner(&self) -> &T {
+    pub const fn inner(&self) -> &T {
         &self.0
     }
 }